HUB-5364 Extract requirement-form visibility, scroll-spy, and event hooks
Split the oversized RequirementForm into a directory with focused hooks so checklist visibility and scroll-spy stay maintainable without changing the public import path.
