+++ app/frontend/components/shared/permit-applications/requirement-form/index.tsx
@@ -7,32 +7,28 @@ import * as R from "ramda"
 import React, { useEffect, useMemo, useRef, useState } from "react"
 import { Trans, useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
-import { useMountStatus } from "../../../hooks/use-mount-status"
-import { IPermitApplication } from "../../../models/permit-application"
-import { EFileUploadAttachmentType, EFlashMessageStatus, EStepCodeType } from "../../../types/enums"
-import { IErrorsBoxData } from "../../../types/types"
-import {
-  getCompletedBlocksFromForm,
-  getRequirementByKey,
-  reconcileSelectedTabIndex,
-  scrollAdjustmentForHiddenBlocks,
-} from "../../../utils/formio-component-traversal"
+import { IPermitApplication } from "../../../../models/permit-application"
+import { EFlashMessageStatus } from "../../../../types/enums"
+import { IErrorsBoxData } from "../../../../types/types"
+import { getRequirementByKey } from "../../../../utils/formio-component-traversal"
 import {
   getEnergyStepCodeMethodFromData,
   singleRequirementFormJson,
   singleRequirementSubmissionData,
-} from "../../../utils/formio-helpers"
-import { downloadFileFromStorage } from "../../../utils/utility-functions"
-import { CompareRequirementsBox } from "../../domains/permit-application/compare-requirements-box"
-import { ErrorsBox } from "../../domains/permit-application/errors-box"
-import { BuilderBottomFloatingButtons } from "../../domains/requirement-template/builder-bottom-floating-buttons"
-import { CustomMessageBox } from "../base/custom-message-box"
-import { SharedSpinner } from "../base/shared-spinner"
-import { Form, defaultOptions } from "../chefs"
-import { ContactModal } from "../contact/contact-modal"
-import { PreviousSubmissionModal } from "../revisions/previous-submission-modal"
-import { PermitApplicationSubmitModal } from "./permit-application-submit-modal"
-import { StepCodeSelectModal } from "./step-code-select-modal"
+} from "../../../../utils/formio-helpers"
+import { CompareRequirementsBox } from "../../../domains/permit-application/compare-requirements-box"
+import { ErrorsBox } from "../../../domains/permit-application/errors-box"
+import { BuilderBottomFloatingButtons } from "../../../domains/requirement-template/builder-bottom-floating-buttons"
+import { CustomMessageBox } from "../../base/custom-message-box"
+import { SharedSpinner } from "../../base/shared-spinner"
+import { Form, defaultOptions } from "../../chefs"
+import { ContactModal } from "../../contact/contact-modal"
+import { PreviousSubmissionModal } from "../../revisions/previous-submission-modal"
+import { PermitApplicationSubmitModal } from "../permit-application-submit-modal"
+import { StepCodeSelectModal } from "../step-code-select-modal"
+import { useBlockScrollSpy } from "./hooks/use-block-scroll-spy"
+import { useChecklistVisibility } from "./hooks/use-checklist-visibility"
+import { useRequirementFormEvents } from "./hooks/use-requirement-form-events"
 
 interface IRequirementFormProps {
   permitApplication?: IPermitApplication
@@ -77,85 +73,44 @@ export const RequirementForm = observer(
       isStepCodeComplete,
     } = permitApplication
 
-    const stepCodeCompletionOptions = { isStepCodeComplete }
-    const previousVisibleBlockKeysRef = useRef<string[]>([])
-    const previousCompletedBlocksRef = useRef<Record<string, boolean>>({})
-    const previousBlockLayoutsRef = useRef<{ key: string; absTop: number; height: number }[]>([])
-    const [visibilityVersion, setVisibilityVersion] = useState(0)
-
-    const captureBlockLayouts = () =>
-      blockClasses.flatMap((className) => {
-        const el = document.getElementsByClassName(className)[0] as HTMLElement | undefined
-        if (!el || el.classList.contains("formio-hidden")) return []
-        const rect = el.getBoundingClientRect()
-        return [
-          {
-            key: className.replace(/^formio-component-/, ""),
-            absTop: rect.top + window.scrollY,
-            height: rect.height,
-          },
-        ]
-      })
-
-    const syncCompletedBlocksFromForm = (root) => {
-      if (!root || !onCompletedBlocksChange) return
-
-      const previousKeys = previousVisibleBlockKeysRef.current
-      const nextCompleted = getCompletedBlocksFromForm(root, stepCodeCompletionOptions)
-      const nextKeys = Object.keys(nextCompleted)
-      const keysChanged = previousKeys.length !== nextKeys.length || previousKeys.some((key, i) => key !== nextKeys[i])
-      const completionChanged =
-        keysChanged || nextKeys.some((key) => previousCompletedBlocksRef.current[key] !== nextCompleted[key])
-
-      if (keysChanged && previousKeys.length > 0) {
-        const hiddenKeys = new Set(previousKeys.filter((key) => !nextKeys.includes(key)))
-        // Bonus: counteract layout collapse when a tall block above the viewport hides.
-        if (hiddenKeys.size > 0) {
-          const adjustment = scrollAdjustmentForHiddenBlocks(
-            previousBlockLayoutsRef.current,
-            hiddenKeys,
-            window.scrollY
-          )
-          if (adjustment > 0) window.scrollBy(0, -adjustment)
-        }
-
-        const nextIndex = reconcileSelectedTabIndex(previousKeys, nextKeys, selectedTabIndex)
-        if (nextIndex !== selectedTabIndex) setSelectedTabIndex(nextIndex)
-      }
-
-      previousVisibleBlockKeysRef.current = nextKeys
-      if (completionChanged) {
-        previousCompletedBlocksRef.current = nextCompleted
-        onCompletedBlocksChange(nextCompleted)
-      }
-      if (keysChanged) {
-        requestAnimationFrame(() => {
-          previousBlockLayoutsRef.current = captureBlockLayouts()
-        })
-        setVisibilityVersion((v) => v + 1)
-      }
-    }
-
-    const syncCompletedBlocksFromFormRef = useRef(syncCompletedBlocksFromForm)
-    syncCompletedBlocksFromFormRef.current = syncCompletedBlocksFromForm
+    const {
+      visibilityVersion,
+      syncCompletedBlocksFromForm,
+      syncCompletedBlocksFromFormRef,
+      previousBlockLayoutsRef,
+      captureBlockLayouts,
+    } = useChecklistVisibility({
+      formRef,
+      blockClasses,
+      selectedTabIndex,
+      setSelectedTabIndex,
+      isStepCodeComplete,
+      onCompletedBlocksChange,
+    })
 
     const shouldShowDiff = permitApplication?.shouldShowApplicationDiff(isEditing)
     const userShouldSeeDiff = permitApplication?.currentUserShouldSeeApplicationDiff
 
     const pastVersion = previousToSelectedSubmissionVersion || previousSubmissionVersion
-    const isMounted = useMountStatus()
     const { t } = useTranslation()
     const navigate = useNavigate()
     const { isOpen, onOpen, onClose } = useDisclosure()
     const boxRef = useRef<HTMLDivElement>(null)
 
-    const [wrapperClickCount, setWrapperClickCount] = useState(0)
+    useBlockScrollSpy({
+      boxRef,
+      formJson,
+      blockClasses,
+      visibilityVersion,
+      setSelectedTabIndex,
+      previousBlockLayoutsRef,
+      captureBlockLayouts,
+    })
+
     const [errorBoxData, setErrorBoxData] = useState<IErrorsBoxData[]>([]) // an array of Labels and links to the component
     const [imminentSubmission, setImminentSubmission] = useState(null)
     const [floatErrorBox, setFloatErrorBox] = useState(false)
     const [hasErrors, setHasErrors] = useState(false)
-    const [autofillContactKey, setAutofillContactKey] = useState(null)
-    const [previousSubmissionKey, setPreviousSubmissionKey] = useState(null)
     const [firstComponentKey, setFirstComponentKey] = useState(null)
     const [isCollapsedAll, setIsCollapsedAllState] = useState(false)
 
@@ -194,6 +149,20 @@ export const RequirementForm = observer(
       onClose: onPreviousSubmissionClose,
     } = useDisclosure()
 
+    const {
+      autofillContactKey,
+      previousSubmissionKey,
+      isStepCodeSelectOpen,
+      setIsStepCodeSelectOpen,
+      stepCodeSelectType,
+      handleSelectExistingStepCode,
+    } = useRequirementFormEvents({
+      permitApplication,
+      triggerSave,
+      onContactsOpen,
+      onPreviousSubmissionOpen,
+    })
+
     const infoBoxData = permitApplication.diffToInfoBoxData
 
     useEffect(() => {
@@ -207,25 +176,9 @@ export const RequirementForm = observer(
       // We don't want to trigger a re-render if the permitApplication itself changes, only if the derived data changes
     }, [displayedSubmissionData])
 
-    // Recompute block completion when the digital Step Code checklist completion flips
-    useEffect(() => {
-      if (formRef.current) syncCompletedBlocksFromFormRef.current(formRef.current)
-    }, [isStepCodeComplete])
-
-    useEffect(() => {
-      // The box observers need to be re-registered whenever a panel is collapsed
-      // This triggers a re-registration whenever the body of the box is clicked
-      // Click listener must be registered this way because formIO prevents bubbling
-
-      const box = boxRef.current
-      const handleClick = () => {
-        setWrapperClickCount((n) => n + 1)
-      }
-      box?.addEventListener("click", handleClick)
-      return () => {
-        box?.removeEventListener("click", handleClick)
-      }
-    }, [])
+    const onScroll = (_event) => {
+      setFloatErrorBox(hasErrors && isFirstComponentNearTopOfView(firstComponentKey))
+    }
 
     useEffect(() => {
       if (hasErrors) {
@@ -237,125 +190,6 @@ export const RequirementForm = observer(
         window.removeEventListener("scroll", onScroll)
       }
     }, [hasErrors])
-
-    useEffect(() => {
-      // This useEffect registers the intersection observers for the panels
-      // It uses a thin line running across the middle of the screen
-      // and detects when this line covers at least a small percentage of the panel
-
-      // Problems with render timing necessitates the use of this isMounted state
-      if (!isMounted) return
-
-      const formComponentNodes = document.querySelectorAll(".formio-component")
-
-      const blockNodes = Array.from(formComponentNodes).filter(
-        (node) =>
-          !node.classList.contains("formio-hidden") &&
-          Array.from(node.classList).some((className) => blockClasses.includes(className))
-      )
-      const viewportHeight = window.innerHeight // Get the viewport height
-      const topValue = -viewportHeight / 2 + 5
-      const bottomValue = -viewportHeight / 2 + 5
-      const rootMarginValue = `${topValue}px 0px ${bottomValue}px 0px`
-      const blockOptions = {
-        rootMargin: rootMarginValue,
-        threshold: 0.0001, // Adjust threshold based on needs
-      }
-
-      const blockObserver = new IntersectionObserver(handleBlockIntersection, blockOptions)
-
-      Object.values(blockNodes).forEach((ref) => {
-        if (ref) {
-          blockObserver.observe(ref)
-        }
-      })
-
-      previousBlockLayoutsRef.current = captureBlockLayouts()
-
-      return () => {
-        blockObserver.disconnect()
-      }
-    }, [formJson, isMounted, window.innerHeight, wrapperClickCount, visibilityVersion])
-
-    const handleOpenStepCodePart3 = async (_event) => {
-      await triggerSave?.()
-      navigate("part-3-step-code")
-    }
-
-    const handleOpenStepCodePart9 = async (_event) => {
-      await triggerSave?.()
-      navigate("part-9-step-code")
-    }
-
-    const handleOpenContactAutofill = async (event) => {
-      setAutofillContactKey(event.detail.key)
-      onContactsOpen()
-    }
-
-    const handleOpenPreviousSubmission = async (event) => {
-      setPreviousSubmissionKey(event.detail.key)
-      onPreviousSubmissionOpen()
-    }
-
-    const [isStepCodeSelectOpen, setIsStepCodeSelectOpen] = useState(false)
-    const [stepCodeSelectType, setStepCodeSelectType] = useState<EStepCodeType>(EStepCodeType.part9StepCode)
-    const handleOpenExistingStepCode = async (event) => {
-      const incoming = event?.detail?.stepCodeType as EStepCodeType
-      setStepCodeSelectType(
-        incoming === EStepCodeType.part3StepCode ? EStepCodeType.part3StepCode : EStepCodeType.part9StepCode
-      )
-      setIsStepCodeSelectOpen(true)
-    }
-
-    const handleSelectExistingStepCode = async (stepCodeId: string) => {
-      await triggerSave?.()
-      // Assign by updating the StepCode's permitApplicationId (belongs_to association)
-      // @ts-ignore method added on model
-      const ok = await permitApplication.assignExistingStepCode(stepCodeId)
-      if (ok) setIsStepCodeSelectOpen(false)
-    }
-
-    const handleDownloadRequirementDocument = async (event) => {
-      downloadFileFromStorage({
-        model: EFileUploadAttachmentType.RequirementDocument,
-        modelId: event.detail.id,
-        filename: event.detail.filename,
-      })
-    }
-
-    const handleOpenResourceLink = (event) => {
-      window.open(event.detail.url, "_blank", "noopener,noreferrer")
-    }
-
-    const handleDownloadResourceDocument = async (event) => {
-      downloadFileFromStorage({
-        model: EFileUploadAttachmentType.ResourceDocument,
-        modelId: event.detail.id,
-        filename: event.detail.filename,
-      })
-    }
-
-    useEffect(() => {
-      document.addEventListener("openStepCode", handleOpenStepCodePart9)
-      document.addEventListener("openStepCodePart3", handleOpenStepCodePart3)
-      document.addEventListener("openAutofillContact", handleOpenContactAutofill)
-      document.addEventListener("openPreviousSubmission", handleOpenPreviousSubmission)
-      document.addEventListener("openExistingStepCode", handleOpenExistingStepCode)
-      document.addEventListener("downloadRequirementDocument", handleDownloadRequirementDocument)
-      document.addEventListener("openResourceLink", handleOpenResourceLink)
-      document.addEventListener("downloadResourceDocument", handleDownloadResourceDocument)
-
-      return () => {
-        document.removeEventListener("openStepCode", handleOpenStepCodePart9)
-        document.removeEventListener("openStepCodePart3", handleOpenStepCodePart3)
-        document.removeEventListener("openAutofillContact", handleOpenContactAutofill)
-        document.removeEventListener("openPreviousSubmission", handleOpenPreviousSubmission)
-        document.removeEventListener("openExistingStepCode", handleOpenExistingStepCode)
-        document.removeEventListener("downloadRequirementDocument", handleDownloadRequirementDocument)
-        document.removeEventListener("openResourceLink", handleOpenResourceLink)
-        document.removeEventListener("downloadResourceDocument", handleDownloadResourceDocument)
-      }
-    }, [])
 
     const setIsCollapsedAll = (isCollapsedAll: boolean) => {
       if (isCollapsedAll) {
@@ -370,20 +204,6 @@ export const RequirementForm = observer(
       errors.map((error) => {
         return { label: error.component.label, id: error.component.id, class: error.component.class }
       })
-
-    function handleBlockIntersection(entries: IntersectionObserverEntry[]) {
-      const entry = entries.filter((en) => en.isIntersecting)[0]
-      if (!entry) return
-
-      // Index among currently visible blocks so sidebar tabs stay aligned after conditionals hide panels.
-      const visibleBlockNodes = Array.from(document.querySelectorAll(".formio-component")).filter(
-        (node) =>
-          !node.classList.contains("formio-hidden") &&
-          Array.from(node.classList).some((className) => blockClasses.includes(className))
-      )
-      const index = visibleBlockNodes.indexOf(entry.target)
-      if (index >= 0) setSelectedTabIndex(index)
-    }
 
     const onFormSubmit = async (submission: any) => {
       // Form.io does not validate the digital Step Code tool (button container, input:false).
@@ -411,9 +231,6 @@ export const RequirementForm = observer(
 
     const onBlur = (containerComponent) => {
       syncCompletedBlocksFromForm(containerComponent.root)
-    }
-    const onScroll = (event) => {
-      setFloatErrorBox(hasErrors && isFirstComponentNearTopOfView(firstComponentKey))
     }
 
     const onChange = (changedEvent) => {
@@ -452,7 +269,7 @@ export const RequirementForm = observer(
         syncCompletedBlocksFromFormRef.current(formRef.current)
       })
 
-      rootComponent.on("submitError", (error) => {
+      rootComponent.on("submitError", (_error) => {
         // when the form attempts to submit but has validation errors, we set a flag to show ErrorBox
         setHasErrors(true)
         setErrorBoxData(mapErrorBoxData(formRef.current.errors))
